% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_veg_classic}
\alias{read_veg_classic}
\title{Read the classic VIC vegetation parameter file.}
\usage{
read_veg_classic(vegfile, rootbands = 3, hasLAI = F)
}
\arguments{
\item{vegfile}{Path of the vegetation parameter file.}

\item{rootbands}{Number of root bands.}

\item{hasLAI}{If the file includes LAI data.}
}
\value{
The vegetation parameter input (A list) for the run of
VIC model in this package.
}
\description{
Read the ASCII vegetation parameter file
of the classic VIC model
}

