% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_params.R
\name{vic_param}
\alias{vic_param}
\title{Get or set global parameters of the VIC model.}
\usage{
vic_param(par, val = NULL)
}
\arguments{
\item{par}{Name of the VIC global parameter.}

\item{val}{Value of the parameter to be set. Will return the
current setting value when not provided.}
}
\description{
Get or set global parameters of the VIC model.
}

